+++ src/textual/app.py
@@ -488,20 +488,32 @@ class App(Generic[ReturnType], DOMNode):
     HORIZONTAL_BREAKPOINTS: ClassVar[list[tuple[int, str]]] | None = []
     """List of horizontal breakpoints for responsive classes.
 
-    This allows for styles to be responsive to the dimensions of the terminal.
-    For instance, you might want to show less information, or fewer columns on a narrow displays -- or more information when the terminal is sized wider than usual.
-    
-    A breakpoint consists of a tuple containing the minimum width where the class should applied, and the name of the class to set.
+    Setting this classvar on `App` or `Screen` will automatically apply TCSS classes to your Screen according to the terminal width.
+    You can then use these classes to adjust the UI to better fit your content to the current terminal dimensions.
 
-    Note that only one class name is set, and you should avoid having more than one breakpoint set for the same size.
+    A breakpoint consists of a tuple containing the minimum width where the class should be applied, and the name of the class to set.
+    Only a single class name may be given, and you should avoid having more than one breakpoint for the same size.
 
-    Example:
-        ```python
-        # Up to 80 cells wide, the app has the class "-normal"
-        # 80 - 119 cells wide, the app has the class "-wide"
-        # 120 cells or wider, the app has the class "-very-wide"
+    Set `HORIZONTAL_BREAKPOINTS` on your app or screen as follows: 
+
+    ```python
+    class MyApp(App):
+        # Up to 80 cells wide, the screen has the class "-normal"
+        # 80 - 119 cells wide, the screen has the class "-wide"
+        # 120 cells or wider, the screen has the class "-very-wide"
         HORIZONTAL_BREAKPOINTS = [(0, "-normal"), (80, "-wide"), (120, "-very-wide")]
-        ```
+    ```
+
+    You can then appliy specfific rules for breakpoints with a rule that targets your screen with the given breakpoint:
+    
+    ```
+    Screen.-wide {
+        # Rules for wide terminals here.
+    }
+    Screen.-very-wide {
+        # Rules for very wide terminals.
+    }
+    ```
     
     """
     VERTICAL_BREAKPOINTS: ClassVar[list[tuple[int, str]]] | None = []
@@ -529,7 +541,7 @@ class App(Generic[ReturnType], DOMNode):
         "dark": lambda app: app.current_theme.dark,
         "light": lambda app: not app.current_theme.dark,
         "inline": lambda app: app.is_inline,
-        "ansi": lambda app: app.ansi_color,
+        "ansi": lambda app: app.native_ansi_color,
         "nocolor": lambda app: app.no_color,
     }
 
@@ -554,7 +566,7 @@ class App(Generic[ReturnType], DOMNode):
     ansi_theme_light = Reactive(ALABASTER, init=False)
     """Maps ANSI colors to hex colors using a Rich TerminalTheme object while using a light theme."""
 
-    ansi_color = Reactive(False)
+    ansi_color: Reactive[bool | None] = Reactive(None)
     """Allow ANSI colors in UI?"""
 
     def __init__(
@@ -562,7 +574,7 @@ class App(Generic[ReturnType], DOMNode):
         driver_class: Type[Driver] | None = None,
         css_path: CSSPathType | None = None,
         watch_css: bool = False,
-        ansi_color: bool = False,
+        ansi_color: bool | None = None,
     ):
         """Create an instance of an app.
 
@@ -574,7 +586,7 @@ class App(Generic[ReturnType], DOMNode):
                 will be loaded in order.
             watch_css: Reload CSS if the files changed. This is set automatically if
                 you are using `textual run` with the `dev` switch.
-            ansi_color: Allow ANSI colors if `True`, or convert ANSI colors to RGB if `False`.
+            ansi_color: Allow ANSI colors if `True`, or convert ANSI colors to RGB if `False`, `None` to use "ansi" parameter from themes.
 
         Raises:
             CssPathError: When the supplied CSS path(s) are an unexpected type.
@@ -596,12 +608,12 @@ class App(Generic[ReturnType], DOMNode):
         )
         self.set_reactive(App.ansi_color, ansi_color)
         self._filters: list[LineFilter] = [
-            ANSIToTruecolor(ansi_theme, enabled=not ansi_color)
+            ANSIToTruecolor(ansi_theme, enabled=not self.native_ansi_color)
         ]
         environ = dict(os.environ)
         self.no_color = environ.pop("NO_COLOR", None) is not None
         if self.no_color:
-            self._filters.append(NoColor() if self.ansi_color else Monochrome())
+            self._filters.append(NoColor() if self.native_ansi_color else Monochrome())
 
         for filter_name in constants.FILTERS.split(","):
             filter = filter_name.lower().strip()
@@ -1095,8 +1107,12 @@ class App(Generic[ReturnType], DOMNode):
             active_message_pump.reset(message_pump_reset_token)
             active_app.reset(app_reset_token)
 
-    def _watch_ansi_color(self, ansi_color: bool) -> None:
+    def _watch_ansi_color(self, ansi_color: bool | None) -> None:
         """Enable or disable the truecolor filter when the reactive changes"""
+        if ansi_color is None:
+            if (theme := self.get_theme(self.theme)) is not None:
+                ansi_color = theme.ansi
+
         for filter in self._filters:
             if isinstance(filter, ANSIToTruecolor):
                 filter.enabled = not ansi_color
@@ -1320,12 +1336,11 @@ class App(Generic[ReturnType], DOMNode):
         Yields:
             [SystemCommand][textual.app.SystemCommand] instances.
         """
-        if not self.ansi_color:
-            yield SystemCommand(
-                "Theme",
-                "Change the current theme",
-                self.action_change_theme,
-            )
+        yield SystemCommand(
+            "Theme",
+            "Change the current theme",
+            self.action_change_theme,
+        )
         yield SystemCommand(
             "Quit",
             "Quit the application as soon as possible",
@@ -1492,9 +1507,14 @@ class App(Generic[ReturnType], DOMNode):
         """
         theme = self.current_theme
         dark = theme.dark
-        self.ansi_color = theme_name == "textual-ansi"
-        self.set_class(dark, "-dark-mode", update=False)
-        self.set_class(not dark, "-light-mode", update=False)
+        # Setting the theme adds class "-theme-<THEME NAME>" to the App
+        classes = {name: False for name in self.classes if name.startswith("-theme-")}
+        classes[f"-theme-{self.current_theme.name}"] = True
+        classes["-dark-mode"] = dark
+        classes["-light-mode"] = not dark
+
+        self.update_classes(classes, update=False)
+
         self._refresh_truecolor_filter(self.ansi_theme)
         self._invalidate_css()
         self.call_next(partial(self.refresh_css, animate=False))
@@ -1527,6 +1547,16 @@ class App(Generic[ReturnType], DOMNode):
             self.ansi_theme_dark if self.current_theme.dark else self.ansi_theme_light
         )
 
+    @property
+    def native_ansi_color(self) -> bool:
+        """Use native ANSI colors?
+
+        This will return `self.current_theme.ansi` if `self.ansi_color` is `None`,
+        otherwise it will return `self.ansi_color`.
+        """
+        ansi = self.current_theme.ansi if self.ansi_color is None else self.ansi_color
+        return ansi
+
     def _refresh_truecolor_filter(self, theme: TerminalTheme) -> None:
         """Update the ANSI to Truecolor filter, if available, with a new theme mapping.
 
@@ -1534,9 +1564,10 @@ class App(Generic[ReturnType], DOMNode):
             theme: The new terminal theme to use for mapping ANSI to truecolor.
         """
         filters = self._filters
+        ansi_color = self.native_ansi_color
         for index, filter in enumerate(filters):
             if isinstance(filter, ANSIToTruecolor):
-                filters[index] = ANSIToTruecolor(theme, enabled=not self.ansi_color)
+                filters[index] = ANSIToTruecolor(theme, enabled=not ansi_color)
                 return
 
     def get_driver_class(self) -> Type[Driver]: