+++ src/textual/visual.py
@@ -10,7 +10,6 @@ from rich.console import Console, ConsoleOptions, RenderableType
 from rich.measure import Measurement
 from rich.protocol import is_renderable, rich_cast
 from rich.segment import Segment
-from rich.style import NULL_STYLE as RICH_NULL_STYLE
 from rich.style import Style as RichStyle
 from rich.text import Text
 
@@ -219,10 +218,8 @@ class Visual(ABC):
 
         selection = widget.text_selection
         if selection is not None:
-            selection_style: Style | None = Style.from_rich_style(
-                widget.screen.get_component_rich_style(
-                    "screen--selection", default=RICH_NULL_STYLE
-                )
+            selection_style = Style.from_styles(
+                widget.screen.get_component_styles("screen--selection")
             )
         else:
             selection_style = None
@@ -238,8 +235,11 @@ class Visual(ABC):
                 selection_style,
             ),
         )
-        if widget.auto_links and not widget.is_container:
-            # TODO: This is suprisingly expensive (why?)
+        if (
+            widget.auto_links
+            and not widget.is_container
+            and not widget.screen._selecting
+        ):
             link_style = widget.link_style
             strips = [strip._apply_link_style(link_style) for strip in strips]