+++ src/textual/widgets/_switch.py
@@ -54,6 +54,20 @@ class Switch(Widget, can_focus=True):
         width: auto;
         pointer: pointer;
         
+        &:ansi {
+            & .switch--slider {
+                color: ansi_red;
+            }
+            &:hover {
+                & > .switch--slider {
+                    color: ansi_bright_red;
+                }
+                &.-on > .switch--slider {
+                    color: ansi_bright_green;
+                }
+            }
+        }
+
         padding: 0 2;
         &.-on .switch--slider {
             color: $success;