+++ src/textual/widgets/_toggle_button.py
@@ -62,6 +62,20 @@ class ToggleButton(Static, can_focus=True):
         text-overflow: ellipsis;
         pointer: pointer;
 
+        &:ansi {
+            background: ansi_default;
+            & > .toggle--button {
+                background: $ansi-background;
+                color: $ansi-foreground;
+                text-style: dim;
+            }
+            &.-on > .toggle--button {
+                color: $accent;
+                background: $ansi-background;
+                text-style: bold not dim;
+            }
+        }
+        
         &.-textual-compact {
             border: none !important;
             padding: 0;